Cleaning up import statements
