+++ core/neo4j/rel_to_neo4j.py
@@ -1,10 +1,11 @@
 """
 Objective: The goal of this script is to create relationships in Neo4j directly from the pipeline using class instances
 """
+import time
 
 from py2neo import Graph
+
 from core.neo4j.nodes_to_neo4j import prep
-import time
 
 # Merge to Neo4j Destop.
 # g = Graph("bolt://localhost:11002", user="neo4j", password="1234")
